backend/requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
google-genai
pyaudio
python-dotenv
